Moved some enums from models
