+++ EVEStandard/Enumerations/ActionEnum.cs
@@ -1,0 +1,16 @@
+﻿namespace EVEStandard.Enumerations
+{
+    public enum ActionEnum
+    {
+        add = 1,
+        assemble = 2,
+        configure = 3,
+        enter_password = 4,
+        Lock = 5,
+        move = 6,
+        repackage = 7,
+        set_name = 8,
+        set_password = 9,
+        unlock = 10
+    }
+}
