+++ EVEStandard/Enumerations/ContactTypeEnum.cs
@@ -1,0 +1,10 @@
+﻿namespace EVEStandard.Enumerations
+{
+    public enum ContactTypeEnum
+    {
+        character = 1,
+        corporation = 2,
+        alliance = 3,
+        faction = 4
+    }
+}
