+++ EVEStandard/Enumerations/GenderEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace EVEStandard.Enumerations
+{
+    public enum GenderEnum
+    {
+        female = 1,
+        male = 2
+    }
+}
