+++ EVEStandard/Enumerations/LocationTypeEnum.cs
@@ -1,0 +1,9 @@
+﻿namespace EVEStandard.Enumerations
+{
+    public enum LocationTypeEnum
+    {
+        station,
+        solar_system,
+        other
+    }
+}
