+++ EVEStandard/Enumerations/PlanetTypeEnum.cs
@@ -1,0 +1,14 @@
+﻿namespace EVEStandard.Enumerations
+{
+    public enum PlanetTypeEnum
+    {
+        temperate = 1,
+        barren = 2,
+        oceanic = 3,
+        ice = 4,
+        gas = 5,
+        lava = 6,
+        storm = 7,
+        plasma = 8
+    }
+}
