+++ EVEStandard/Enumerations/RoleTypeEnum.cs
@@ -1,0 +1,14 @@
+﻿namespace EVEStandard.Enumerations
+{
+    public enum RoleTypeEnum
+    {
+        grantable_roles,
+        grantable_roles_at_base,
+        grantable_roles_at_hq,
+        grantable_roles_at_other,
+        roles,
+        roles_at_base,
+        roles_at_hq,
+        roles_at_other
+    }
+}
