+++ EVEStandard/Models/Asset.cs
@@ -8,17 +8,6 @@
 
     public class Asset : ModelBase<Asset>
     {
-        /// <summary>
-        ///     location_type string
-        /// </summary>
-        /// <value>location_type string</value>
-        public enum LocationTypeEnum
-        {
-            station,
-            solar_system,
-            other
-        }
-
         [JsonProperty("location_flag"), JsonConverter(typeof(StringEnumConverter))] public LocationFlag LocationFlag { get; set; }
 
         [JsonProperty("location_id")] public long LocationId { get; set; }
